fix(policy): split exit_decision.py under the 800-line architecture cap
exit_decision.py grew to 824 lines after ADR-064's scan/release resolvers,
tripping the policy package's production-file cap. Move
resolve_scan_exit_decision/resolve_release_exit_decision (and their
_dominant_decision/_DOMINANT_FIELD support) into a new sibling module,
exit_decision_precedence.py, and update the flat back-compat shim plus the
migration test to resolve a shimmed name across both real modules. Also
fixes a stale ADR-064 cross-reference that still described stage 1a as
"not yet implemented".
